+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - **JEI Improvements**: Ability to drag items from JEI when creating manifests
 - **Shopping Cart Improvements**: Holding the `Shift` and `Control` keys down while pressing the `+` or `-` buttons now applies a multiplier
   - `Shift` is for 64 items, while `Control` is for 10
+- **Sell All**: By shift-clicking the shopping cart button in the admin shop, all items that you are able to sell will be added to the cart
 
 ## 1.2.2-1.6.2